# libevent-based http server