--- a/depends/packages/libevent.mk
+++ b/depends/packages/libevent.mk
@@ -3,6 +3,11 @@ $(package)_version=2.0.22
 $(package)_download_path=https://github.com/libevent/libevent/releases/download/release-2.0.22-stable
 $(package)_file_name=$(package)-$($(package)_version)-stable.tar.gz
 $(package)_sha256_hash=71c2c49f0adadacfdbe6332a372c38cf9c8b7895bb73dabeaa53cdcc1d4e1fa3
+$(package)_patches=reuseaddr.patch
+
+define $(package)_preprocess_cmds
+  patch -p1 < $($(package)_patch_dir)/reuseaddr.patch
+endef
 
 define $(package)_set_vars
   $(package)_config_opts=--disable-shared --disable-openssl --disable-libevent-regress