… contrib/debian/examples/bitcoin.conf
… doc/build-unix.md
… doc/release-notes.md
… src/Makefile.qt.include
… src/Makefile.qttest.include
… src/httpserver.cpp
… src/init.cpp
… src/init.h
… src/qt/bitcoin.cpp
… src/qt/rpcconsole.cpp
… src/qt/rpcconsole.h
… src/qt/test/test_main.cpp
… src/rest.cpp
… src/rpcprotocol.cpp
… src/rpcprotocol.h
… src/rpcserver.cpp
… src/rpcserver.h
… src/test/rpc_tests.cpp
… src/util.cpp
… src/util.h