@theuni 2015-01-20
Very nice! Before looking over the work itself, I wanted to be sure that libevent was viable for all of our build targets.

See here for the build-system work. This should be enough to get Travis passing, I'd think:
https://github.com/theuni/bitcoin/commits/5677