@luke-jr 2015-01-20
Hm, nothing special needed for longpolling?