@laanwj 2015-01-20
@luke-jr I don't think so. The current implementation should work. Of course it would be more optimal to release the worker thread while longpolling, and change "new block" it into an event-trigger, but I leave that as a challenge for later.