@laanwj 2015-01-20
The fail in "32-bit + dash" is strange

    FAIL: qt/test/test_bitcoin-qt

I'm not sure how this can be affected at all, but I'll check.

On Win32/64 it still tries to link against libevent_pthread. IIRC there is no specific thread library for windows, `evthread_use_windows_threads` is part of the core library there.