@laanwj 2015-01-20
It's easy to miss those sneaky qt unit tests. Windows passes now!

That leaves the `32-bit + dash` case, which is not an intermittent issue
```
FAIL! : PaymentServerTests::paymentServerTests() Compared values are not the same
Actual (merchant):
Expected (QString("testmerchant.org")): testmerchant.org
Loc: [qt/test/paymentservertests.cpp(84)]
```
... no clue how this happens yet. The Qt tests don't use the RPC mechanism. My gut feeling is some interaction with OpenSSL which, absent verification, is now an indirect dependency through Qt? Not sure, and why it only happens in this test is a mystery to me.