@theuni 2015-01-20
@laanwj Yes, it has some interaction with qt:
- ./configure --with-gui=qt4: fine.
- ./configure --with-gui=qt5: fine.
- make -C depends; ./configure --prefix=\`pwd\`/depends/x86_64-unknown-linux-gnu: busted
- make -C depends USE_LINUX_STATIC_QT5=1;  ./configure --prefix=\`pwd\`/depends/x86_64-unknown-linux-gnu: fine.