@jonasschnelli 2015-01-20
tested gitian build.
Binaries to test: https://builds.jonasschnelli.ch/pulls/5677/