@laanwj 2015-01-23
Now that the code is stable it is time for some benchmarking.
I found a nice scriptable framework for HTTP benchmarking, [wrk](https://github.com/wg/wrk). Some results.
These benchmarks were taken at the default settings (4 worker threads, 16 depth work queue).

### GET request to invalid URL
These are handled by evhttp itself, so this is the baseline.
```
$  ./wrk -t12 -c15 -d10s http://127.0.0.1:18332/inv
  Thread Stats   Avg      Stdev     Max   +/- Stdev
    Latency   375.53us   66.86us   6.31ms   93.03%
    Req/Sec     2.67k   121.28     3.11k    80.98%
  303293 requests in 10.00s, 36.73MB read
  Non-2xx or 3xx responses:
    404: 303293
Requests/sec:  30343.08
Transfer/sec:      3.68MB
```
As expected, we get 404s.
### GET request to /

These are dispatched to a worker thread. Some more latency is expected. In the worker thread these error out early, as GET is not a valid method for JSON RPC.
```
$ ./wrk -t12 -c15 -d10s http://127.0.0.1:18332/
  Thread Stats   Avg      Stdev     Max   +/- Stdev
    Latency   472.96us   44.12us   1.28ms   93.31%
    Req/Sec     2.14k   127.02     2.33k    82.90%
  245025 requests in 10.00s, 41.59MB read
  Non-2xx or 3xx responses:
    405: 245025
Requests/sec:  24514.30
Transfer/sec:      4.16MB
```
As expected: lots of 405 (invalid method for URL) results.
### RPC getgenerate requests
Post `getgenerate` requests. This is an extremely cheap RPC call.
Script:
```lua
wrk.method = "POST"
wrk.body   = "{\"method\":\"getgenerate\",\"params\":[],\"id\":1}\n"
wrk.headers["Content-Type"] = "application/json"
wrk.headers["Authorization"] = "Basic XXX"
```
```
$ ./wrk -t12 -c15 -d10s -s getgenerate.lua http://127.0.0.1:18332/
  Thread Stats   Avg      Stdev     Max   +/- Stdev
    Latency   571.28us   59.77us   2.09ms   91.89%
    Req/Sec     1.82k   110.56     2.00k    80.32%
  204481 requests in 9.99s, 28.28MB read
  Non-2xx or 3xx responses:
Requests/sec:  20463.77
Transfer/sec:      2.83MB
```
All responses succesful.
### RPC getinfo requests
Post `getinfo` requests. This is a more expensive RPC call.
```lua
wrk.method = "POST"
wrk.body   = "{\"method\":\"getinfo\",\"params\":[],\"id\":1}\n"
wrk.headers["Content-Type"] = "application/json"
wrk.headers["Authorization"] = "Basic XXX
```
```
$ ./wrk -t12 -c15 -d10s -s getinfo.lua http://127.0.0.1:18332/
  Thread Stats   Avg      Stdev     Max   +/- Stdev
    Latency     1.69ms  572.87us  15.39ms   95.21%
    Req/Sec   626.46     81.57     1.00k    84.26%
  71245 requests in 9.99s, 37.17MB read
  Non-2xx or 3xx responses:
Requests/sec:   7128.48
Transfer/sec:      3.72MB
```
Although the performance goes down, no errors happen. The maximum queue depth is never reached.

### RPC requests w/ invalid authentication
Post `getinfo` requests with invalid authentication. This will trigger a 250ms delay, and thus we can trigger worker queue-full conditions with enough threads.
```lua
wrk.method = "POST"
wrk.body   = "{\"method\":\"getinfo\",\"params\":[],\"id\":1}\n"
wrk.headers["Content-Type"] = "application/json"
wrk.headers["Authorization"] = "Basic YYY
```
```
$ ./wrk -t12 -c15 -d10s -s getinfo_unauth.lua http://127.0.0.1:18332/
Thread Stats   Avg      Stdev     Max   +/- Stdev
    Latency   751.03ms  323.38us 752.15ms   81.82%
    Req/Sec     1.05      0.65     2.00     58.33%
  156 requests in 10.00s, 19.80KB read
  Non-2xx or 3xx responses:
    401: 156
Requests/sec:     15.60
Transfer/sec:      1.98KB
```
Increasing the load further, we can exceed the work queue depth:
```
$ ./wrk -t12 -c45 -d10s -s getinfo_unauth.lua http://127.0.0.1:18332/
  Thread Stats   Avg      Stdev     Max   +/- Stdev
    Latency   204.35ms  462.08ms   1.25s    83.72%
    Req/Sec     2.42k     1.53k    6.00k    75.64%
  277072 requests in 10.00s, 43.59MB read
  Non-2xx or 3xx responses:
    401: 156
    500: **276916**
Requests/sec:  27706.71
Transfer/sec:      4.36MB
```

Looks good. Nothing unexpected, it's clear the evhttp is not the bottleneck, and the work queue works as expected. Will try this with the old asio-based HTTP server shortly.