@laanwj 2015-01-23
## Old http server
Same steps as above, repeated with the old server as of commit 944c256.

### GET request to invalid URL
```
$  ./wrk -t12 -c15 -d10s http://127.0.0.1:18332/inv
  Thread Stats   Avg      Stdev     Max   +/- Stdev
    Latency   502.17us   94.60us   2.65ms   74.77%
    Req/Sec     1.89k   119.76     2.33k    83.38%
  214864 requests in 9.99s, 37.50MB read
  Responses:
    404: 214864
Requests/sec:  21502.28
Transfer/sec:      3.75MB
```
### GET request to /

```
$ ./wrk -t12 -c15 -d10s http://127.0.0.1:18332/
  Thread Stats   Avg      Stdev     Max   +/- Stdev
    Latency   502.58us   92.30us   2.07ms   73.39%
    Req/Sec     1.89k   118.97     2.22k    83.36%
  214269 requests in 9.99s, 103.40MB read
  Responses:
    401: 214269
Requests/sec:  21447.93
Transfer/sec:     10.35MB
```

### RPC getgenerate requests
```
$ ./wrk -t12 -c15 -d10s -s getgenerate.lua http://127.0.0.1:18332/
  Thread Stats   Avg      Stdev     Max   +/- Stdev
    Latency    93.68us   24.84us   1.27ms   89.23%
    Req/Sec    10.06k   616.30    11.78k    64.49%
  380211 requests in 10.00s, 78.32MB read
  Socket errors: connect 0, read 0, write 0, timeout 39
  Responses:
    200: 380211
Requests/sec:  38006.24
Transfer/sec:      7.83MB
```
### RPC getinfo requests
```
$ ./wrk -t12 -c15 -d10s -s getinfo.lua http://127.0.0.1:18332/
Running 10s test @ http://127.0.0.1:18332/
  12 threads and 15 connections
  Thread Stats   Avg      Stdev     Max   +/- Stdev
    Latency   544.15us  187.74us   4.03ms   79.81%
    Req/Sec     1.89k   188.50     2.90k    63.09%
  71480 requests in 10.00s, 42.13MB read
  Socket errors: connect 0, read 0, write 0, timeout 39
  Responses:
    200: 71480
Requests/sec:   7144.83
Transfer/sec:      4.21MB
```
### RPC requests w/ invalid authentication
```
$ ./wrk -t12 -c15 -d10s -s getinfo_unauth.lua http://127.0.0.1:18332/
  12 threads and 15 connections
  Thread Stats   Avg      Stdev     Max   +/- Stdev
    Latency   751.75ms  161.83us 752.03ms   73.91%
    Req/Sec     0.94      0.66     2.00     56.52%
  156 requests in 10.00s, 77.09KB read
  Responses:
    401: 156
Requests/sec:     15.60
Transfer/sec:      7.71KB
```
```
$ ./wrk -t12 -c45 -d10s -s getinfo_unauth.lua http://127.0.0.1:18332/
Running 10s test @ http://127.0.0.1:18332/
  12 threads and 45 connections
  Thread Stats   Avg      Stdev     Max   +/- Stdev
    Latency     2.19s   309.47ms   2.26s    95.24%
    Req/Sec     0.90      1.56     6.00     87.30%
  156 requests in 10.01s, 77.09KB read
  Socket errors: connect 0, read 0, write 0, timeout 24
  Responses:
    401: 156
Requests/sec:     15.59
Transfer/sec:      7.70KB
```

## Summary

```
Testcase                      Old (r/s) New (r/s)
================================================
GET request to invalid URL    21502     30343
GET request to /              21448     24514
RPC getgenerate requests      38006*    20463
RPC getinfo requests           7144*     7128
RPC requests w/ invalid auth     16*       16

* with timeout errors.
```

- The new server wins on the base http requests front. Even to /, which are dispatched to the worker threads.

- The old server is currently much faster with `getgenerate` requests. I am curious why. Also: how can getgenerate requests be faster than simple `GET /`'s? I suspect a measurement error that has to do with the timeouts.

- `getinfo` requests are, as expected, ~ the same speed. Processing overhead dominates.

- Same for unauthenticated requests. The requests take 250ms to handle so the number seen is exactly as expected.